Remove unneeded code.
... oops! I made some and forgot to remove! my bad!
also made sure clippy is happy.

+++ src/channel_row.rs
@@ -74,7 +74,7 @@ impl DewChannelRow {
     }
     fn set_subs(&self, subs: f32) {
         self.imp().subs.set_text(
-            &(util::format_semi_engineering(subs as f32) + " subscribers"),
+            &(util::format_semi_engineering(subs) + " subscribers"),
         );
     }
 }
